delay dealled

+++ osptest/source/server.cpp
@@ -3,6 +3,7 @@
 
 #define MAKEESTATE(state,event) ((u32)(event<< 4 +state))
 
+
 CSApp g_cCSApp;
 
 s8 buffer[BUFFER_SIZE];
@@ -102,14 +103,16 @@ void CSInstance::MsgProcessInit(){
 
 
       //common Instance
-        RegMsgProFun(MAKEESTATE(IDLE_STATE,FILE_NAME_SEND),&CSInstance::FileNameSend,&m_tCmdChain);
+        RegMsgProFun(MAKEESTATE(IDLE_STATE,FILE_RECEIVE_UPLOAD),&CSInstance::FileReceiveUpload,&m_tCmdChain);
+        RegMsgProFun(MAKEESTATE(RUNNING_STATE,FILE_RECEIVE_REMOVE),&CSInstance::ReceiveRemove,&m_tCmdChain);
+        RegMsgProFun(MAKEESTATE(RUNNING_STATE,FILE_RECEIVE_CANCEL),&CSInstance::ReceiveCancel,&m_tCmdChain);
+
         RegMsgProFun(MAKEESTATE(RUNNING_STATE,FILE_UPLOAD),&CSInstance::FileUpload,&m_tCmdChain);
         RegMsgProFun(MAKEESTATE(RUNNING_STATE,FILE_FINISH),&CSInstance::FileFinish,&m_tCmdChain);
         RegMsgProFun(MAKEESTATE(RUNNING_STATE,FILE_CANCEL),&CSInstance::FileCancel,&m_tCmdChain);
         RegMsgProFun(MAKEESTATE(RUNNING_STATE,FILE_REMOVE),&CSInstance::FileRemove,&m_tCmdChain);
-        RegMsgProFun(MAKEESTATE(RUNNING_STATE,FILE_RECEIVE_REMOVE),&CSInstance::ReceiveRemove,&m_tCmdChain);
-        RegMsgProFun(MAKEESTATE(RUNNING_STATE,FILE_RECEIVE_CANCEL),&CSInstance::ReceiveCancel,&m_tCmdChain);
         RegMsgProFun(MAKEESTATE(RUNNING_STATE,FILE_GO_ON),&CSInstance::FileGoOn,&m_tCmdChain);
+        RegMsgProFun(MAKEESTATE(RUNNING_STATE,FILE_STABLE_REMOVE),&CSInstance::FileStableRemove,&m_tCmdChain);
 }
 
 void CSInstance::NodeChainEnd(){
@@ -221,41 +224,85 @@ void CSInstance::SignOut(CMessage* const pcMsg){
       OspPrintf(1,1,"sign out\n");
 }
 
-void CSInstance::FileNameSend(CMessage* const pMsg){
+void CSInstance::FileReceiveUpload(CMessage* const pMsg){
 
+#ifdef _LINUX_
+        struct flock fl;
+#endif
      if(!pMsg->content || pMsg->length <= 0){
-             OspLog(LOG_LVL_ERROR,"[FileNameSend]file name is NULL\n");
+             OspLog(LOG_LVL_ERROR,"[FileReceiveUpload]file name is NULL\n");
              //TODO:通知客户端
              return;
      }
      if(strlen((LPCSTR)pMsg->content)+1 != pMsg->length){
-         OspLog(LOG_LVL_ERROR,"[FileNameSend]file name error\n");
+         OspLog(LOG_LVL_ERROR,"[FileReceiveUpload]file name error\n");
              //TODO:通知客户端
              return;
      }
      if(pMsg->length > MAX_FILE_NAME_LENGTH-1 ||
                      pMsg->length <= 0 || !pMsg->content){
-             OspLog(LOG_LVL_ERROR,"[InstanceEntry] file name error\n");
+             OspLog(LOG_LVL_ERROR,"[FileReceiveUpload] file name error\n");
              //TODO:通知客户端
              return;
      }
 
+     if(access((LPCSTR)pMsg->content,F_OK) != -1){
+             OspLog(SYS_LOG_LEVEL,"[FileReceiveUpload]file exists\n");
+             //TODO:通知客户端文件已存在
+#if 0
+             if(!file_cover){
+                 return;
+             }
+#endif
+     }
      memcpy(file_name_path,pMsg->content,pMsg->length);
-     if(!(file = fopen((LPCSTR)pMsg->content,"wb"))){
-             //TODO:通知客户端
-             OspLog(LOG_LVL_ERROR,"file open error\n");
-             printf("open file error\n");
+#ifdef _LINUX_
+     if(INVALID_FILEHANDLE == (file = open((LPCSTR)pMsg->content,O_WRONLY | O_CREAT))){
+             //TODO:通知客户端,get the errno
+             OspLog(LOG_LVL_ERROR,"[FileReceiveUpload]file open error\n");
+             perror("[FileReceiveUpload]open file error\n");
              return;
      }
+     fl.l_type = F_WRLCK;
+     fl.l_whence = SEEK_SET;
+     fl.l_start = 0;
+     fl.l_len = 0; //specify 0:lock all bytes to the end of file,no matter how large
+                 //the file grows
+     if(fcntl(file,F_SETLK,&fl) == -1){
+             if(errno == EACCES || errno == EAGAIN){//already locked
+                     OspLog(LOG_LVL_ERROR,"[FileReceiveUpload]Already locked by anther process\n");
+                     if(OSP_OK != post(pMsg->srcid,FILE_LOCKED,NULL
+                           ,0,pMsg->srcnode)){
+                             OspPrintf(1,0,"[FileReceiveUpload]post back failed\n");
+                             printf("[FileReceiveUpload]post back failed\n");
+                     }
+                     return;
+             }else{
+                     OspLog(LOG_LVL_ERROR,"[FileReceiveUpload]check file locking unexpected error\n");
+#if 0
+                     //通知客户端
+                     if(OSP_OK != post(pMsg->srcid,FILE_LOCKING_ERROR,NULL
+                           ,0,pMsg->srcnode)){
+                             OspPrintf(1,0,"post back failed\n");
+                             printf("post back failed\n");
+                     }
+#endif
+                     return;
 
-     if(OSP_OK != post(pMsg->srcid,FILE_NAME_ACK,NULL
+             }
+     }
+#elif defined _MSC_VER
+#endif
+
+     if(OSP_OK != post(pMsg->srcid,FILE_RECEIVE_UPLOAD_ACK,NULL
            ,0,pMsg->srcnode)){
-             OspPrintf(1,0,"post back failed\n");
-             printf("post back failed\n");
+             OspPrintf(1,0,"[FileReceiveUpload]post back failed\n");
+             printf("[FileReceiveUpload]post back failed\n");
      }
 
      NextState(RUNNING_STATE);
-     printf("file name send\n");
+     emFileStatus = STATUS_RECEIVE_UPLOAD;
+     printf("[FileReceiveUpload]file name send\n");
 }
 
 
@@ -263,15 +310,29 @@ void CSInstance::FileUpload(CMessage* const pMsg){
 
      //TODO:增加缓冲
      if(pMsg->content && pMsg->length > 0){
+#if 0
              if(fwrite(pMsg->content,1,sizeof(s8)*pMsg->length,file)
                              != pMsg->length || ferror(file)){
                      OspLog(LOG_LVL_ERROR,"file upload write error\n");
                      //TODO:通知客户端
                      return;
              };
+#endif
+#ifdef _LINUX_
+             if(-1 == write(file,pMsg->content,sizeof(s8)*pMsg->length)){
+                     //TODO:print errno
+                     OspLog(LOG_LVL_ERROR,"file upload write error\n");
+                     //TODO:通知客户端
+                     return;
+             }
+#elif defined _MSC_VER
+#endif
      }
      //TODO:需要增加返回信息，为客户端文件传送进度显示做依据。
      printf("get files\n");
+     if(emFileStatus == STATUS_RECEIVE_UPLOAD){
+        emFileStatus = STATUS_UPLOADING;
+     }
      if(OSP_OK != post(pMsg->srcid,FILE_UPLOAD_ACK,&emFileStatus
            ,sizeof(emFileStatus),pMsg->srcnode)){
              OspPrintf(1,0,"post back failed\n");
@@ -283,17 +344,57 @@ void CSInstance::FileUpload(CMessage* const pMsg){
 void CSInstance::ReceiveCancel(CMessage* const pMsg){
 
         OspLog(SYS_LOG_LEVEL,"[ReceiveCancel]receive cancel msg\n");
-        emFileStatus = RECEIVE_CANCEL;
+        emFileStatus = STATUS_RECEIVE_CANCEL;
 }
 
 void CSInstance::ReceiveRemove(CMessage* const pMsg){
 
-
         OspLog(SYS_LOG_LEVEL,"[ReceiveRemove]receive remove msg\n");
-        emFileStatus = RECEIVE_REMOVE;
+        emFileStatus = STATUS_RECEIVE_REMOVE;
 }
 
 void CSInstance::FileGoOn(CMessage* const pMsg){
+
+#ifdef _LINUX_
+
+     struct flock fl;
+     if(INVALID_FILEHANDLE == (file = open((LPCSTR)file_name_path,O_WRONLY | O_APPEND))){
+             //TODO:通知客户端
+             OspLog(LOG_LVL_ERROR,"file open error\n");
+             perror("open file error\n");
+             return;
+     }
+     fl.l_type = F_WRLCK;
+     fl.l_whence = SEEK_SET;
+     fl.l_start = 0;
+     fl.l_len = 0; //specify 0:lock all bytes to the end of file,no matter how large
+                 //the file grows
+     if(fcntl(file,F_SETLK,&fl) == -1){
+             if(errno == EACCES || errno == EAGAIN){//already locked
+                     OspLog(LOG_LVL_ERROR,"Already locked by anther process\n");
+                     if(OSP_OK != post(pMsg->srcid,FILE_LOCKED,NULL
+                           ,0,pMsg->srcnode)){
+                             OspPrintf(1,0,"post back failed\n");
+                             printf("post back failed\n");
+                     }
+                     return;
+             }else{
+                     perror("FileGoOn\n");
+                     OspLog(LOG_LVL_ERROR,"[FileGoOn]check file locking unexpected error\n");
+#if 0
+                     //通知客户端
+                     if(OSP_OK != post(pMsg->srcid,FILE_LOCKING_ERROR,NULL
+                           ,0,pMsg->srcnode)){
+                             OspPrintf(1,0,"post back failed\n");
+                             printf("post back failed\n");
+                     }
+#endif
+                     return;
+
+             }
+     }
+#elif defined _MSC_VER
+#endif
 
         if(OSP_OK != post(pMsg->srcid,FILE_GO_ON_ACK,NULL
               ,0,pMsg->srcnode)){
@@ -301,33 +402,33 @@ void CSInstance::FileGoOn(CMessage* const pMsg){
                 printf("post back failed\n");
                 return;
         }
+#if 0
         if(!(file = fopen((LPCSTR)file_name_path,"ab"))){
                 //TODO:通知客户端
                 OspLog(LOG_LVL_ERROR,"file open error\n");
                 printf("open file error\n");
                 return;
         }
-        emFileStatus = GO_ON_SEND;
+#endif
+
+        emFileStatus = STATUS_UPLOADING;
 }
 
 void CSInstance::FileRemove(CMessage* const pMsg){
 
-        if(emFileStatus < CANCELED){
-                if(fclose(file) == 0){
-                        OspLog(SYS_LOG_LEVEL,"[FileRemove]file closed\n");
-
-                }else{
-                        OspLog(LOG_LVL_ERROR,"[FileRemove]file close failed\n");
-                        //TODO：通知客户端
-                        return;
-                }
-        }
 #ifdef _LINUX_
+        if(-1 == close(file)){//record locks removed
+                OspLog(LOG_LVL_ERROR,"[FileRemove]file close failed\n");
+                //get the errno
+                //TODO：通知客户端
+                return;
+        }
+        OspLog(SYS_LOG_LEVEL,"[FileRemove]file closed\n");
         if(unlink((LPCSTR)file_name_path) == 0){
                 OspLog(SYS_LOG_LEVEL,"[FileRemove]file removed\n");
-                emFileStatus = REMOVED;
+                emFileStatus = STATUS_REMOVED;
                 if(OSP_OK != post(pMsg->srcid,FILE_REMOVE_ACK
-                                    ,&emFileStatus,sizeof(emFileStatus),pMsg->srcnode)){
+                                    ,NULL,0,pMsg->srcnode)){
                         OspPrintf(1,0,"[FileRemove]post back failed\n");
                         printf("[FileRemove]post back failed\n");
                 }
@@ -335,12 +436,36 @@ void CSInstance::FileRemove(CMessage* const pMsg){
                 OspLog(LOG_LVL_ERROR,"[FileRemove]file remove failed\n");
                 //TODO：通知客户端
         }
+
+#elif defined _MSC_VER_
 #endif
+        NextState(IDLE_STATE);
+}
+
+void CSInstance::FileStableRemove(CMessage* const pMsg){
+
+#ifdef _LINUX_
+        if(unlink((LPCSTR)file_name_path) == 0){
+                OspLog(SYS_LOG_LEVEL,"[FileRemove]file removed\n");
+                emFileStatus = STATUS_REMOVED;
+                if(OSP_OK != post(pMsg->srcid,FILE_STABLE_REMOVE_ACK
+                                    ,NULL,0,pMsg->srcnode)){
+                        OspPrintf(1,0,"[FileRemove]post back failed\n");
+                        printf("[FileRemove]post back failed\n");
+                }
+        }else{
+                OspLog(LOG_LVL_ERROR,"[FileRemove]file remove failed\n");
+                //TODO：通知客户端
+        }
+#elif defined _MSC_VER_
+#endif
+        NextState(IDLE_STATE);
 }
 
 void CSInstance::FileCancel(CMessage* const pMsg){
 
         //保存为临时文件，
+#if 0
         if(fclose(file) == 0){
                 OspLog(SYS_LOG_LEVEL,"[FileCancel]file closed\n");
 
@@ -349,7 +474,19 @@ void CSInstance::FileCancel(CMessage* const pMsg){
                 //TODO：通知客户端
                 return;
         }
-        emFileStatus = CANCELED;
+#endif
+#ifdef _LINUX_
+     if(-1 == close(file)){//record locks removed
+             OspLog(LOG_LVL_ERROR,"[FileCancel]file close failed\n");
+             //get the errno
+             //TODO：通知客户端
+             return;
+     }
+     OspLog(SYS_LOG_LEVEL,"[FileCancel]file closed\n");
+#elif defined _MSC_VER_
+#endif
+
+        emFileStatus = STATUS_CANCELLED;
         if(OSP_OK != post(pMsg->srcid,FILE_CANCEL_ACK
                             ,&emFileStatus,sizeof(emFileStatus),pMsg->srcnode)){
                 OspPrintf(1,0,"[FileCancel]post back failed\n");
@@ -360,6 +497,7 @@ void CSInstance::FileCancel(CMessage* const pMsg){
 void CSInstance::FileFinish(CMessage* const pMsg){
 
      if(pMsg->content && pMsg->length > 0){
+#if 0
              if(fwrite(pMsg->content,1,sizeof(s8)*pMsg->length,file)
                              != pMsg->length || ferror(file)){
                      OspLog(LOG_LVL_ERROR,"file upload write error\n");
@@ -374,8 +512,22 @@ void CSInstance::FileFinish(CMessage* const pMsg){
                      //TODO:通知客户端
                      return;
              }
+#endif
+#ifdef _LINUX_
+             if(-1 == write(file,pMsg->content,sizeof(s8)*pMsg->length)){
+                     //TODO:print errno
+                     OspLog(LOG_LVL_ERROR,"[FileFinish] write error\n");
+                     NextState(IDLE_STATE);
+                     //TODO:close file
+                     //TODO:通知客户端
+                     return;
+             }
+#elif defined _MSC_VER
+#endif
+
      }
      OspLog(SYS_LOG_LEVEL,"file finished\n");
+#if 0
      if(fclose(file) == 0){
              OspLog(SYS_LOG_LEVEL,"[FileRemove]file closed\n");
 
@@ -383,11 +535,23 @@ void CSInstance::FileFinish(CMessage* const pMsg){
              OspLog(LOG_LVL_ERROR,"[FileRemove]file close failed\n");
              //TODO：通知客户端
      }
+#endif
+#ifdef _LINUX_
+     if(-1 == close(file)){//record locks removed
+             OspLog(LOG_LVL_ERROR,"[FileFinish]file close failed\n");
+             //get the errno
+             //TODO：通知客户端
+             return;
+     }
+     OspLog(SYS_LOG_LEVEL,"[FileFinish]file closed\n");
+#elif defined _MSC_VER_
+#endif
      if(OSP_OK != post(pMsg->srcid,FILE_FINISH_ACK,NULL
            ,0,pMsg->srcnode)){
              OspPrintf(1,0,"post back failed\n");
              printf("post back failed\n");
      }
      NextState(IDLE_STATE);
-     emFileStatus = FINISHED;
+     emFileStatus = STATUS_FINISHED;
 }
+
